Adding a monthly ranking (#1938)
This PR add coder of the month rank that was mentioned in #1790 

The changes consist of:
1. Adding a new column `rank` in coder of the month table
This rank will be used to show the the ranking of coder of the month. By specifying rank default value as 1.
2. Calculate rank inside `calculateCoderOfTheMonth`
Formerly this function only calculates 1 person to be the coder of the month. This function will calculate 100 users as coder of the month.
3. Vuefication `coderofthemonth.tpl`
Vuefication was done because we need to create a new UI for showing user the coder of the month rank and there are some issue about vuefication smarty template.