IT WORKS!
Turns out the timeouts were due to the common chunks not being
completely extracted, so every single entry was including the whole
contents of omegaup.js and all its dependencies (including the entire
Vue runtime).

This change uses the chunk settings that are compatible with Webpack 4,
which includes the need to extract the common chunks into something that
is _not_ an entrypoint.