[Feature request]External link will open in new tab in help section (#6505)
The flow of site is in single tab it is okay also. As each step is linked.
But for help section there are two list item one for Tutorial and one is guide.
These should be open in new tab by default so site flow remain maintained As these are external references.