Aligned the navigation bar buttons on the right side for mobile users… (#6527)
For mobile users, the navigation bar buttons appeared on the left side while the navbar button was on the right which causes problem to many. 

In this pull request, I moved the navigation bar buttons on the right side for mobile users. It doesn't affect the  web view. 

I have used `align-items` tag instead of `float` as stated in the previous Pull Request. 

In the video below you can see the the PR doesn't affect the web view and make the mobile view much better. 
![Screenshot from 2022-04-14 23-09-24](https://user-images.githubusercontent.com/63846699/163443840-f5a06c5f-a102-474f-b945-06a64c33c052.png)

https://user-images.githubusercontent.com/63846699/162439035-35c1022d-0e38-4741-a9e3-997e08fab14e.mp4

Fixes:#5934