Fix bug "Endless redirects"
This line of code made the following bug:

When a user visits /libinteractive/index.php the intended redirection should be /libinteractive/en

But because there was an error in this line of code "header("Location: {$location}{$preferred}/");"

This code redirects the user to the URL {$location}{$preferred}/. However, the value of {$location} is the current request URI, which is /libinteractive/index.php. This means that the user is being redirected to the URL /libinteractive/index.php/{$preferred}/. However, this URL is the same as the original URL, so the user is redirected back to /libinteractive/index.php, which then redirects the user back to /libinteractive/index.php/{$preferred}/, and so on.

Thus creating a infinite loop and recursive function.

The new code:
"header("Location: /libinteractive/{$preferred}/");"

This should take the user to the desired language page.