Added an event that copies content of editor into modal when mounted (#7482)
# Description

I triggered an event manually on **mounted** method of `CodeView`, this
event is added using javascript window addEventListener method. This
event updates an object to **have the same code and language of that of
editor**, then I use that object to mount(init) the modal with the same
data as the editor taking into consideration that these changes should
be made visible to the parent of `CodeView`.
Here is a video of what I mean:


https://github.com/omegaup/omegaup/assets/59091536/729fce84-552d-47cf-881c-8d56c786b530

I have been trying a more `Vue` friendly approach using event bus or
`Vuex` store but I couldn't get them to connect properly. This -I
believe- is because the ephemeral editor is embedded where the
`CodeView` is on one window while the editor is on another.


I submit this as a draft because I would like to know if I have made any
subtle mistakes with naming or otherwise, your feedback is appreciated!

Fixes: #7450

# Comments

Add any comments for the reviewers (e.g., indicating the beginning of
the revision, something where the reviewer needs to pay special
attention, etc.). If there are no extra comments this section can be
deleted.

# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [x] The tests were executed and all of them passed.

---------

Co-authored-by: Hugo Dueñas <heduenas@users.noreply.github.com>