Add filterBySignedUp() method to ContestListv2.vue (#7485)
# Description

triggerd the filterBySignedUp when the component is mounted if the url
containes the "#participating" hhash, so that the filtering could take
place and user can see the list of filters contest in which he/she has
enrolled in

Fixes: #7480 

# Comments

Add any comments for the reviewers (e.g., indicating the beginning of
the revision, something where the reviewer needs to pay special
attention, etc.). If there are no extra comments this section can be
deleted.

# Checklist:

- [ ] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [ ] The tests were executed and all of them passed.
- [ ] If you are creating a feature, the new tests were added.
- [ ] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.

---------

Co-authored-by: Hugo Dueñas <heduenas@users.noreply.github.com>