Changed id of messages (#7521)
# Description

Changed the id of review messages to start from 1 instead of 0. This
gives the correct boolean value which is used for disabling the submit
review button.

![Screenshot 2024-03-24
180411](https://github.com/omegaup/omegaup/assets/67920373/e53f213a-1aac-4fe2-afc2-a8299135f271)


Fixes: #7515 

# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [ ] If you are creating a feature, the new tests were added.
- [ ] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.

---------

Co-authored-by: Juan Pablo Gómez Aguilar <gomez_aguilar_juan_pablo@hotmail.com>
Co-authored-by: Hugo Dueñas <heduenas@users.noreply.github.com>