Added remaining languagues to monaco editor, matched them with correct file extensions and editor modes (#7488)
Added remaining languages to `util.js` and updated
`index-light.html`options so that you can pick any of these languages.
Now the languages of monaco editor matches that of the modal. Made sure
to add correct file extensions and modes for each language.
Before (live version):

![l6](https://github.com/omegaup/omegaup/assets/59091536/7cd44d52-8d1d-42d0-b00e-55fafded8faf)
After:

![image](https://github.com/omegaup/omegaup/assets/59091536/71cef282-15af-464b-a631-1569ead8b46f)

![image](https://github.com/omegaup/omegaup/assets/59091536/42e3847b-cbbe-4f9b-bbc5-d5b83dbbfaba)

I wanted to refactor all attributes of a language like this:
```
// TO DO: later iterations, refactor language logic using this object.
// this object should carry all information needed about each language
export const languageInfo = {
  c: { extension: 'c', name: 'C11 (gcc 10.3)', modelMapping: 'cpp' },
  'c11-gcc': { extension: 'c', name: 'C11 (gcc 10.3)', modelMapping: 'cpp' },
  ....
  }
 ```
And then export it throughout the project, but I think that will take some time and debugging. I believe this pr fixes the issue for now.


# Description

Please include a description of the changes you have made. The text you put in this section will be used as the **commit message** when this PR is merged.

If you modify the **User Interface**, please include a screenshot or a video.

Fixes: #7483

# Comments

Add any comments for the reviewers (e.g., indicating the beginning of the revision, something where the reviewer needs to pay special attention, etc.). If  there are no extra comments this section can be deleted.

# Checklist:

- [x] The code follows the [coding guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines) of omegaUp.
- [x] The tests were executed and all of them passed.

---------

Co-authored-by: Hugo Dueñas <heduenas@users.noreply.github.com>