Remove scroll CSS styles from ContestListv2.vue (#7518)
# Description

After reading some of the previous commits this was also asked to be
fixed previously, but wasn't done correctly. I have removed the styles
for scroll-content, hence enablin the height to be auto.


![image](https://github.com/omegaup/omegaup/assets/111554475/3927480a-24bc-46e4-b728-6c2b62725aba)

here it is displaying 10 contest yet no scrolls

Fixes: #7516 

# Comments

Add any comments for the reviewers (e.g., indicating the beginning of
the revision, something where the reviewer needs to pay special
attention, etc.). If there are no extra comments this section can be
deleted.

# Checklist:

- [ ] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [ ] The tests were executed and all of them passed.
- [ ] If you are creating a feature, the new tests were added.
- [ ] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.