Arreglando la propiedad de los problemas que se pueden borrar (#7556)
# Description

In the PR #7476, the flag `can_be_removed` was added to the `List` and
`AdminList` APIs.

However, it had a bug: the flag turned on when the problems actually
couldn't be removed, and it turned off for problems that could be
removed.

Part of: #7452 

# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [x] If you are creating a feature, the new tests were added.
- [x] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.