Fixed Submisison Heading (#7559)
# Description

It used to say Submission on the last 24 hours, changed it to Submission
in the last 24 hours, the correct one.

Fixes: #(issue)

# Comments

Add any comments for the reviewers (e.g., indicating the beginning of
the revision, something where the reviewer needs to pay special
attention, etc.). If there are no extra comments this section can be
deleted.

# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [x] If you are creating a feature, the new tests were added.
- [x] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.

---------

Co-authored-by: Hugo Dueñas <heduenas@users.noreply.github.com>