Webpack cache busting grader directory (#7539)
# Description

Problem:
Browser was caching `ephemeral.js`. So, any changes made on
`ephemeral.js` will not be reflected as the browser will **return cached
version** of `ephemeral.js` not updated one.

![grader-cache-1](https://github.com/omegaup/omegaup/assets/59091536/91485577-0701-4e75-a465-dcf803feba33)
Why is it doing that for `ephemeral.js`?
I think it's because it's a large file (9 MB) so the browser decides to
cache that file, but I'm not sure.
Solution:
Implement cache busting for `webpack.config-grader` file. Put simply,
cache busting is a way to tell the browser **not to cache certain
files/assets when we don't want them to**, like in this scenario. What I
did was I moved `index.html` and `index-light.html` to a separate
directory called `templates`. Then, I use the `HtmlWebpackPlugin` to
init the `index.html` and `index-light.html` with actual hashed
`ephemeral.js` script. Now it works properly.

Fixes: #7524

# Comments
1) Each time we update `ephemeral.js` and new hashed script is created
in `dist` directory. We need to find a way to `clean` previous hashed
when creating new one.
2) Put resulting `index.html` and `index-light.html` created by the
plugin in `.gitignore` file.
3) Webpack will `defer` the script, there is no `async` scriptLoading. I
believe deferring the script fixes #7493
# Checklist:

- [ ] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [ ] The tests were executed and all of them passed.
- [ ] If you are creating a feature, the new tests were added.
- [ ] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.

---------

Co-authored-by: Hugo Dueñas <heduenas@users.noreply.github.com>