Agregando Submissions::apiList para obtener envíos por páginas (#7555)
In the PR #7553 will be implemented the infinite scroll for the latest
submissions, but it needs to create `api.Submissions.list` in order to
get a defined number of rows when user scrolls down the Submissions List
page.

Part of: #7532

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [x] If you are creating a feature, the new tests were added.
- [x] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.

Co-authored-by: Hugo Dueñas <heduenas@users.noreply.github.com>