Add the problem statement tab (#7607)
This PR adds the `problem statement` tab to the creator page..

Fixes: #7587



https://github.com/omegaup/omegaup/assets/64671908/049bccc2-f6d9-41b6-b801-961cacec625f



# Comments:
Following the unofficial CDP, we will update the markdown content and
the store only on the `Save/Trigger` button. The content will be stored
in the store which will be used to generate the zip file.

This PR
- [x] adds the problem statement tab to the creator.
- [ ] Writes the `jest` test for it.

# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [x] If you are creating a feature, the new tests were added.
- [x] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.

---------

Signed-off-by: Aritra Majumder <aritramajumder8438@gmail.com>
Co-authored-by: Carlos Córdova Sáenz <carlosab1802@gmail.com>
Co-authored-by: Hugo Dueñas <heduenas@users.noreply.github.com>