Add `rebase.yml` for auto rebase (#7766)
# Description

I have successfully created and tested the automatic `rebase` feature. I
have observed that online sources say that workflows must be present in
the `default` branch of the repository to be available for execution.
Maybe we can merge to `main` and test it again just to be sure. Here is
video manually testing `\rebase`:


https://github.com/user-attachments/assets/f1edcaa2-d910-4b7e-a891-66182246628c

The action can be triggered by users who are members of the
organization. There is an expression allowing `CI` workflow to run after
`rebase` is finished, although this expression is currently set to
false.

Fixes: #7706

# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [ ] If you are creating a feature, the new tests were added.

---------

Co-authored-by: Hugo Dueñas <heduenas@users.noreply.github.com>