Added `Jest` mocks for upcoming `monaco-editor` for upcoming `Ephemeral` component (#7757)
# Description
Added mock object for `editor.create`, `createDiffEditor`, ... for
`Jest` testing with `monaco-editor`. These mocks are necessary for when
the `Ephemeral` component is successfully implemented as that is how it
will recognize the `monaco-editor` package.

Fixes: #7755 

# Comments

# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [x] If you are creating a feature, the new tests were added.
- [x] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.