Complete `Ephemeral.vue` component for problems page (#7782)
# Description

I have added logic for both submitting and running a problem. The
component is now ready to be used instead of the `ephemeral.ts` script
for the problems page case.


https://github.com/user-attachments/assets/c361f42c-0def-4e2b-a7fb-388beef84134

Here is also a video showing that it works with a contest:


https://github.com/user-attachments/assets/351010dd-78bc-4e68-bd21-2f2ddb7784f1


There will be another separate pull request for full screen IDE.
Fixes: #6214, fixes: #7761, fixes: #7749

# Comments
There are multiple files that are not used any longer like
`config-grader`. These will be addressed when completing the component
for full screen IDE.
I modified some existing tests to suit this component.
# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [x] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.

---------

Co-authored-by: Hugo Dueñas <heduenas@users.noreply.github.com>