Add zip `download` and `upload` functionality to `Ephemeral` component (#7799)
# Description

This is a "minified" view of our full IDE, a page for the component will
be addressed in https://github.com/omegaup/omegaup/pull/7794. Download
and upload buttons in action, notice that the icons appear properly:


https://github.com/user-attachments/assets/105266d9-aca3-4852-81de-b98daf855169



Fixes: #7792, fixes: #7816

# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [ ] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.

---------

Co-authored-by: Juan Pablo Gómez Aguilar <gomez_aguilar_juan_pablo@hotmail.com>