Mostrando incondicionalmente todos las categorías de nivel de problemas (#7912)
# Description

In this change, all cards containing the tag categories related to
problem levels are displayed.

This is necessary to make changes to the problem collections UI, as it
will show all the elements that this view can contain.

Part of: #7911

# Comments

When a collection doesn't contain any problems, the "View problems"
button will be disabled.

# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [ ] If you are creating a feature, the new tests were added.
- [x] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.