File Name Change for Windows Users (Final) (#8043)
File Name Change for Windows Users
Description
This PR updates the filename
“…EPRECADO_-Cómo-empezar-a-desarrollar-(entorno-manual).md” to ensure
compatibility for Windows users.

Please provide a brief description of the change. The text in this
section will be used as the commit message when this PR is merged.

If the User Interface has been modified, include a relevant screenshot
or video.

Fixes: #(issue)

Comments
Add any notes for the reviewers, such as specific areas to focus on or
important details about the revision. If there are no additional
comments, this section can be removed.

Checklist
The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
 All tests have been executed and passed.
If this change introduces a new feature, relevant tests have been added.
If the PR is large (> 200 lines), it has been split into multiple Pull
Requests. It is recommended to create one PR for controllers + PHPUnit
tests, and another for UI changes + Jest/Cypress tests.