Add user profile information display in contest intro component (#8004)
# Description

Added the suggested solution of showing the fields which are not filled.


![image](https://github.com/user-attachments/assets/e3f1ea6b-1771-45cb-83da-68d7e934324b)


Fixes: #7780 7780


# Checklist:

- [ ] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [ ] The tests were executed and all of them passed.
- [ ] If you are creating a feature, the new tests were added.
- [ ] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.

---------

Co-authored-by: Hugo Dueñas <heduenas@users.noreply.github.com>