Fix bug about enrolled contest page (#8019)
# Description

Clicking on the Contest enrolled link from the drop down is giving 400
error. This behavior was because the filter was participating and must
be a signedup.

Fixes: #8018

# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [ ] If you are creating a feature, the new tests were added.
- [ ] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.

---------

Co-authored-by: Juan Pablo Gómez Aguilar <gomez_aguilar_juan_pablo@hotmail.com>
Co-authored-by: Hugo Dueñas <heduenas@users.noreply.github.com>