Added the news table in schema (#8047)
# Description

I have added a table News in the schema.sql and a migration file to
enter sample news for the news carousel as mentioned.


Part of: #6916 

# Comments

I was unable to edit the dao_schema.sql file and hence couldn't generate
the dao and vo files for the table. Please approve this pr asap so that
i can start working on this issue.


# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [ ] If you are creating a feature, the new tests were added.
- [x] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.