Add DAO and VO files for managing carousel items (#8158)
# Description

The DAO and VO files are required to interact with the table in mysql
We have added the table in a PR which have already been merged - #8047
In this PR we are adding the DAO and VO files for the table

- Updated the dao_schema.sql file to include the `Carousel_Items` table.
 - Ran the script `./stuff/update-dao.sh` 
 - Updated the DAO and VO files for the new table `Carousel_Items`


Part of: #6916 

# Comments

Add any comments for the reviewers (e.g., indicating the beginning of
the revision, something where the reviewer needs to pay special
attention, etc.). If there are no extra comments this section can be
deleted.

# Checklist:

- [ ] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/wiki/Coding-guidelines)
of omegaUp.
- [ ] The tests were executed and all of them passed.
- [ ] If you are creating a feature, the new tests were added.
- [ ] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.