Documentation translation to English: Subpart-> Conversion of REST-API.md (#8213)
# Description

Created an English translation of the `REST-API.md` file. The content
has also been reviewed and updated to ensure it reflects the current
implementation and standards of the omegaUp REST API.

Fixes: #8200

# Comments

This is part of the "Documentation translation to English" project. The
file has been translated with attention to clarity and structure,
ensuring it's suitable for external developers. No UI changes were made,
so no screenshots are necessary.

# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/blob/main/frontend/www/docs/Coding-guidelines.md)
of omegaUp.
- [ ] The tests were executed and all of them passed.
- [ ] If you are creating a feature, the new tests were added.
- [ ] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.

---------

Co-authored-by: Juan Pablo Gómez Aguilar <gomez_aguilar_juan_pablo@hotmail.com>