Fix typo in column name for Carousel_Items table (#8242)
# Description

During the review process of this PR
https://github.com/omegaup/omegaup/pull/8047
One of the reviewers suggested a commit that we change the primary key
name, but it had a minor typo in it ```corousel_item_id``` but it should
have been ```carousel_item_id```. Mistakenly this got pushed and merged
as well.
This PR fixes this typo.

Fixes: #8243 

# Comments

There would be a follow up PR after this is merged in order to fix the
typo in the DAO and VO files as well.

# Checklist:

- [ ] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/blob/main/frontend/www/docs/Coding-guidelines.md)
of omegaUp.
- [ ] The tests were executed and all of them passed.
- [ ] If you are creating a feature, the new tests were added.
- [ ] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.

---------

Co-authored-by: Juan Pablo Gómez Aguilar <gomez_aguilar_juan_pablo@hotmail.com>