Make assignment alias optional (#8356)
# Description

This PR allows teachers to run the `TA script` for all the assignments
if the `assignment alias` is not given.

This PR is based on #8333 and should be merged before it.
- [x] #8333

# Checklist:

- [x] The code follows the [coding
guidelines](https://github.com/omegaup/omegaup/blob/main/frontend/www/docs/Coding-guidelines.md)
of omegaUp.
- [x] The tests were executed and all of them passed.
- [x] If you are creating a feature, the new tests were added.
- [x] If the change is large (> 200 lines), this PR was split into
various Pull Requests. It's preferred to create one PR for changes in
controllers + unit tests in PHPUnit, and then another Pull Request for
UI + tests in Jest, Cypress or both.

---------

Co-authored-by: Aritra <aritramajumder8438@gmail.com>