added label for PRs automation for the associated change it is associ… (#8596)
I have tried to make sure not to miss any paths. Now, if a PR is being
raised, the associated label with the change will be added
automatically. If there are some tags which need to be there, let me
know. Currently, I have focused on the existing tags.

Fixes #8558

---------

Co-authored-by: Juan Pablo Gómez Aguilar <gomez_aguilar_juan_pablo@hotmail.com>